restructure backend
restructured backend and added a dummy endpoint to access our sql db.

backend structure:
1. main.go:  creates mux, calls registerRoutes, and wraps with cors handler.
2. routers/routers.go: registers routes and wraps in authentication middleware.
3. controllers/profile_controller.go: deals with logic of accessing profile table in db.
4. repositories/profile_repository.go: deals with data transfer with sql db.
5. middlewares/auth.go: implements auth middleware
