+++ backend/go.mod
@@ -4,6 +4,7 @@ go 1.23.2
 
 require (
 	github.com/golang-jwt/jwt/v4 v4.5.1
+	github.com/gorilla/mux v1.8.1
 	github.com/joho/godotenv v1.5.1
 	github.com/rs/cors v1.11.1
 )
